Use boto3 config mechanism if AWS keys are not in the ini
Boto3 will look in several locations when searching for credentials. The mechanism in which Boto3 looks for credentials is to search through a list of possible locations and stop as soon as it finds credentials. In the current implementation, this mechanism is not used as the keys are set in the code. With this change, if the creentials are not in the ini file then the code will use boto3 credential search.

The order in which Boto3 searches for credentials is detailed on https://boto3.amazonaws.com/v1/documentation/api/latest/guide/credentials.html